ir_seekrt_sim/requirements.txt
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
pyyaml>=6.0
tifffile>=2023.1.1
opencv-python>=4.7.0